Introduce DiscoverHudsonPluginsTask
This breaks out the Plugin-Class population into its own task that
generates a dedicated manifest file. Rather than having a subclass
of Manifest be responsible for scanning the classesDirs and
populating a field, we'll move toward generating separate manifests and
combining them just before the jar and jpi are created.

This is part of the work to better model inputs of the manifest and
support the new configuration cache.

JENKINS-58205